// web/src/api-access/generated/models/assignmentDefinitionResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Unified.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface AssignmentDefinitionResponse {
  id?: number;
  locationId?: number;
  name?: string;
  description?: string;
  assignmentCategoryTypeId?: number;
  assignmentCategoryTypeDescription?: string;
  assignmentSubCategoryTypeId?: number;
  assignmentSubCategoryTypeDescription?: string;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  defaultStartTime?: string | null;
  /** @nullable */
  defaultEndTime?: string | null;
  defaultCapacity?: number;
  effectiveDateUtc?: string;
  /** @nullable */
  expiryDateUtc?: string | null;
}